fix(redis/connection): explicit type conversion of DB parameter into int

+++ lib/redis/connection.go
@@ -10,7 +10,7 @@ import (
 var client = redis.NewClient(&redis.Options{
 	Addr:     utils.RedisConfig["host"].(string) + utils.RedisConfig["port"].(string),
 	Password: utils.RedisConfig["password"].(string),
-	DB:       utils.RedisConfig["DB"].(int),
+	DB:       int(utils.RedisConfig["DB"].(float64)),
 })
 
 func init() {
